python-task/requirements.txt
uvicorn
fastapi[all]
aiopg
aiomysql
#ormar[all]
aerich
tortoise-orm[all]
load_dotenv
asyncpg
aiodns
aiohttp[speedups]
httpx
trio
pytest
pandas
openpyxl
xlrd
requests
aiohttp
plotly
dash
dash_bootstrap_components
